Fixed code so nothing is in MainActivity.java or activity_main.xml.
Changed xml file to start on CalendarConnectionActivity.java.
