+++ app/src/main/res/layout/activity_main.xml
@@ -7,71 +7,6 @@
     android:tooltipText="Name of Events"
     tools:context=".MainActivity">
 
-    <EditText
-        android:id="@+id/dateEditText"
-        android:layout_width="262dp"
-        android:layout_height="85dp"
-        android:ems="10"
-        android:hint="Date"
-        android:inputType="date"
-        android:tooltipText="Date"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.543"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.329" />
-
-    <EditText
-        android:id="@+id/descEditText"
-        android:layout_width="264dp"
-        android:layout_height="79dp"
-        android:ems="10"
-        android:hint="Description"
-        android:inputType="text"
-        android:tooltipText="Description"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.551"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.179" />
-
-    <Button
-        android:id="@+id/addEventButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="156dp"
-        android:text="Add Event"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <EditText
-        android:id="@+id/locationEditText"
-        android:layout_width="253dp"
-        android:layout_height="80dp"
-        android:ems="10"
-        android:hint="Location"
-        android:inputType="text"
-        android:tooltipText="Location"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <EditText
-        android:id="@+id/titleEditText"
-        android:layout_width="265dp"
-        android:layout_height="72dp"
-        android:layout_marginTop="24dp"
-        android:ems="10"
-        android:hint="Title of Event"
-        android:inputType="text"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.554"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
